Added an API (kind of) to be imported as a python module and used in other projects
Added an API (kind of) to be imported as a python module and used in other projects. I also updated the main.py a little and cleaned the look of the code. And finally, I removed the useragents file and just put it in as a variable in the main.py file because it wasn't that long.